worker/requirements.txt
pytest==5.3.5
requests==2.22.0
Flask==1.1.1
fake_useragent==0.1.11
python-dotenv==0.13.0
gunicorn==20.0.4
